+++ docs/source/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+jsonyx 2.2.0 (unreleased)
+-------------------------
+
+.. todo:: Add release date
+
+- Added ``cache_keys`` to :class:`jsonyx.Decoder`, :func:`jsonyx.load`,
+  :func:`jsonyx.loads` and :func:`jsonyx.read`
+
 jsonyx 2.1.0 (Mar 30, 2025)
 ---------------------------